+++ js/loaders/loadObjects.js
@@ -1,4 +1,4 @@
-import { objectTypes } from "../viz/constants/objects.js";
+import { objectTypes } from "../viz/constants/objectTypes.js";
 import { supportedEDM4hepTypes } from "../../model/datatypes.js";
 
 export function getSupportedEDM4hepTypes(schemaVersion) {
