+++ js/state/load-event.js
@@ -1,6 +1,6 @@
 import { loadObjects } from "../loaders/load.js";
 import { copyObject } from "../lib/copy.js";
-import { objectTypes } from "../viz/constants/objects.js";
+import { objectTypes } from "../viz/constants/objectTypes.js";
 import { drawView } from "../views/views.js";
 import { getContainer } from "../viz/draw/app.js";
 import { getFileData } from "./file.js";
