+++ js/viz/constants/linkTypes.js
