+++ js/viz/constants/objectTypes.js
