+++ js/viz/lib/getName.js
@@ -1,7 +1,7 @@
-import { mappings } from "../constants/particles.js";
+import { particlesMap } from "../constants/particleMappings.js";
 
 export function getName(pdg) {
-  const particle = mappings.get(String(pdg));
+  const particle = particlesMap.get(String(pdg));
 
   if (particle !== undefined) {
     // console.log("Name: " + particle);
